demos/imagetagger/requirements/development.txt
-e .

mypy
flake8
pytest
pytest-cov
pytest-aiohttp
numpy
keras
Pillow
markdown2
dataclasses
tensorflow
